requirements.txt
langgraph
langchain
langchain-core
pydantic
openai
python-dotenv